Invert GetY() of controllers where necessary as "up" is -Y not +Y (#1122)
On XboxControllers, moving the sticks upwards causes a negative
change on the y-axis value, and vice versa. This means that driving
a motor directly off the y-axis will result in inverted controls
and cause difficulty (or damage, especially IRL). This resolves the
issue by negating the y-axis in calls to the Joystick.